Deprecate `fieldstream` (and `basic_fieldstream`). (#617)
Deprecate field streams.

There are several reasons for this:

First, of course, I doubt many projects use them.  On top of that they
don't add all that much value: it'd be easy enough to do the same things
_ad hoc._

And the C++ streams library is a bit whacky.  It needs an overhaul.
Time to take a step back and decouple from details that may soon change.

Implementing stream classes is complicated, poorly documented, prone to
mistakes, hard in maintenance.  And _using_ streams has similar
problems, especially when it comes to error handling and unexpected
conditions.  It's easy to get wrong.

Finally, besides pointless operator overloading it also adds pointless
inheritance.  It's just not carrying that weight.