Iterate `pqxx::array`. (#619)
Further to the work on #590.

Implementing simple array iteration. And when I say "simple," I really mean the simplest I could do: storage-order iteration, going through all dimensions.

The alternative would be to define something like a multidimensional subrange, and iterate over those. In the two-dimensional case, that would be like iterating a `result` to get `row` elements, and then iterating each `row` to get the `field` elements.  But would it really be all that useful?  It's a lot of code overhead and cognitive load, both on the libpqxx side and on the application side.  I don't want all that complexity again — streaming queries have gotten us away from that for query results and it's been wonderful.  Plus, in practice, all you may want is to iterate over a one-dimensional array.  Or you may just want to visit all elements and not care at all about the order.

It may become easier once we have `std::mdspan`.  Depending on the conventions that form, we may have a problem on our hands with `size()`: is that going to be the size including all elements, or is it just the size along the outer of the dimensions?  The latter seems almost arbitrary.  We'll have to cross that bridge when we come to it.