Make more `zview` constructots `noexcept`.
This is a silly problem to have.  In a way it's good.  Sometimes gcc
warns (and in my development builds, warnings are errors!) that some
`zview` constructors throw no errors, and would I perhaps like to slap
a `noexcept` on them?

A _fine_ idea, were it not for the facts that (1) `pqxx::zview` extends
`std::string_view`, and (2) the corresponding `string_view` constructor
does not have the `noexcept`.  So... can I make that promise myself?

Turns out I can.  Looking at gcc's `string_view` implementation I see
that it does declare those constructors as `noexcept`.  All I had to do
was declare my constructors as "`noexcept` provided that `string_view`
is okay with it.")