`stream_query`: like `stream_from` but stronger-typed.
It's becoming clear that `stream_from` was too weakly typed.  Why would
anyone want to read different rows of the same query result as different
column types?

This is just step 1: a more or less functionally equivalent replacement
of `stream_from`.  The only optimisation so far is replacing a
`std::vector` with a `std::array`.  Now to find some more optimisations.
For instance, there is no need to tiptoe around the possibility of rows
having different numbers of fields.  That's just an error now.

The plan is to move this into the `pqxx::internal` namespace, and make
`transaction_base::stream()` the only way for client code to access it.