Clean up a bit more.
Eliminated the `fields` array — unescape one field at a time, and parse
immediately after, so we never need to keep track of more than one field
at a time.

This also simplifies the folding magic for "iterating" the fields.
Truth be told I never really got how the indexing worked anyway, when it
came to the template trick for the index sequence.  That's all gone now.
I just fold over the field types, and that's that.

Along the way, some awkward small-scale code duplication went out the
window as well.  It feels a bit more sensible to me now.

One thing I'm not _quite_ sure about is the way I mixed two styles of
in-out parameters: at a lower level I use a static member function that
returns modified versions of its arguments, but then at a higher level I
need to pass the values by reference to propagate the side effects in my
folding expression.

Modifying pass-by-reference parameters is ugly, but also, there's always
the worry that the optimiser may no longer be able to figure out what's
going on.  In which case, it may synchronise the values unnecessarily
often, or it may miss rescheduling opportunities for loads and stores,
and so on.