Speed up central streaming loop a bit more.
A COPY line uses (by default, but that's what we stick to) tabs as the
separator between fields, and it ends in a newline.  But we don't really
need to know that that character is the end of the line.  Replacing it
with a tab as a sentinel simplifies the central loop, since it can now
expect every field to end in a tab.