Check for errors after poll/select/WSAPoll. (#655)
Check for errors after poll/select/WSAPoll.

This was never urgent, because if waiting for a socket fails, the
operation which the caller is waiting to perform will normally also
fail.  (And if not... why report an error really?)

But, it's probably nicer to check for errors and report them.