Fix `std::functional` problems. (#677)
Thanks @KayEss for pointing out that I shouldn't be using a `std::function` as a deleter for a `std::shared_ptr` or `std::unique_ptr`.  Using ordinary function pointers instead.

I also worked around a bunch of gcc compile errors when I combined...
* The address sanitizer or "asan" (`-fsanitize=address`).
* Warnings as errors (`-Werror`).
* Uninitialized analyzer (`-Wanalyzer-use-of-uninitialized-value`).

The combination of audit mode (`--enable-audit` in the `configure` script) and maintainer mode (`--enable-maintainer-mode`
in the `configure` script) got us into exactly that combination.

It would seem that there's a problem with gcc.  It doesn't happen in clang.  I worked around it by replacing the one regex in the code with a manual parsing loop like I used to have.