Fix "double free" crash when building as shared.
Fixes #681.

WHen building libpqxx as a shared library on Debian with gcc 12, any
program that included libpqxx headers (with a few exceptions) would
crash after `main()` returned.

There was an error message:

_free(): double free detected in tcache 2_

The problem turned out to be triggered by the inlining of any
specialisation of `pqxx::type_name`, as we did in the macro
`PQXX_DECLARE_ENUM_CONVERSION`.  There's a header that uses that macro,
thus triggering the problem.