Fix "double free" crash when building as shared. (#683)
Fixes #681.

When building libpqxx as a shared library with gcc 12, any program that included libpqxx headers (with a few exceptions) would crash after `main()` returned.

There was an error message:

_free(): double free detected in tcache 2_

The problem turned out to be triggered by the inlining of any specialisation of `pqxx::type_name`, as we did in the macro `PQXX_DECLARE_ENUM_CONVERSION`.  There's a header that uses that macro, thus triggering the problem.

I had to make enum `type_name` a `std::string_view`.  This pains me.  It changes the type of the variable.  But gcc 12 (in a
shared-library build) won't let me get away with an `inline std::string` and clang won't let me get away with a none-`inline` `std::string`.

A `std::string_view` is easier — trivial destructor, no allocations.