Update `config.guess`. (#689)
Really grateful for #688, but there's one problem: it's a lot of work to
review that entire script!  There's some clear improvements, but also a
lot of more or less arbitrary changes.

Yes, yes, I know — the script comes from the GNU site, which we trust
all the time anyway.  But how do I _know_ that it came from the GNU
site?

So here's a lazy shortcut: I download an update of known GNU provenance,
I merge that into the repo, and then hopefully the PR will resolve to a
tiny remaining diff.  Or if we're lucky, none at all.