Forbid conversion of certain types.
This is really just an attempt to produce slighlty more helpful errors
when clients try to convert types that libpqxx deliberately does not
support, such as `char`.  We're talking about conversions that leave
too much to the imagination: a `char` can be signed integral value,
or an unsigned integral value, or a byte of text in some encoding.