Replace some config tests with C++20 feature macros. (#709)
Replace some config tests with C++20 feature macros.

Takes some tedious checks out of the configure stage and replaces them with use of C++20 feature test macros.

Some configure checks will remain, e.g. when we want more fine-grained information than the standard macros (or the compiler's implementation thereof) can offer.

I'm introducing a new header defining my own macros for this. I think it's perfectly reasonable for a libpqxx user to want to compile with options to turn use of an undefined macro into an error. But the C++20 feature test macros normally rely on undefined macros (when used in a preprocessor conditional) evaluating as false. Our own macros are safe to use even with these stricter compiler options. They will also help guard me against typos, which are a frequent problem on this otherwise wonderful laptop.

Expect a similar change for libpq feature tests soon. That library has introduced its own feature test macros. But that work will have to go into its own header, which we will not want to include from the libpqxx headers. By design, libpqxx does not expose any libpq symbols.

It'd be tempting to replace more of the checks right away. But the feature test macros are new in C++20. So in a C++17 compiler, the macro may not be present even for a feature that the compiler does support. This problem will go away with libpqxx 8.0, which will require core C++20 support as a baseline.