Add move constructor to `stream_to`.
Should help with #706, though we may need move assignment as well.

Finally biting the bullet and adding move construction to
`transaction_focus`, which requires dealing with that backpointer from
the transaction back to the focus object.

I'll try move assignment next.  The obvious complication there is that
`m_trans` will have to become a pointer.