Skip pointless compiler option checks. (#710)
In "maintainer mode," with a gcc-compatible compiler, the configure
script tries to enable lots of warning options.  Some of these have been
present in gcc and clang forever, so we just add them blindly.  Others
it tries out first, to see whether the compiler accepts them.

A few of these are now old enough that I'll just enable them.  Shaves a
fraction of a second off the configure step.