Introduce `std::source_location`. (#702)
Introduce `std::source_location`.

Allow exceptions to track source locations.  This is still pretty
rudimentary.

One thing I'm not happy about is the way different exception classes
have their own `source_location` members that are not related to each
other.  This means that if you want to access that field, you have to
catch a fairly specific exception type.

Also of course, it's not great that we can't assume that the compiler
actually supports this feature.  The testing helpers could be much nicer
if they could rely on `source_location` — we might even get by without
some of the preprocessor macros.