Permit conversion to string of `std::string_view`.
Fixes #728.

I've been reluctant to implement this because it could pose a larger
runtime cost than the caller might expect.  We need to copy the whole
underlying string into the available buffer just to ensure termination.

But as the bug shows, users may need the conversion in order to pass
arrays of `string_view` as statement parameters.