Run `ruff`.  Also, `pip install` Python linters. (#762)
Apparently Ubuntu has a "snap" for `ruff`, but it doesn't seem to be
working in 23.10.  So, just `pip install` both ruff and pyflakes if they
are not installed.

It gets a tiny bit complicated with pyflakes, because the Ubuntu package seems to call it `pyflakes3` and the pip package calls it `pyflakes`.  To avoid more complication, I took the risk of setting up an unnecessary virtualenv in one scenario.