Parameterised versions of `query()` etc.
Fixes #646.

The new functions do not have "params" in their names.  I'm hoping to
get to a future where parameterised calls and regular calls look alike,
except that the parameterised calls take an extra `params` argument.

Perhaps in a later stage after that, we can pass parameter packs so
the parameters don't have to be wrapped up in a `params`.  But that's
for later; it complicates the situation with `std::source_location`
parameters which I also want to add, in 8.0.