Some minor tweaks for `char_traits` work. (#792)
* Some minor tweaks for `char_traits` work.

Patches up some very minor details after merging @tambry's great contribution in #751.

After this I'll run the automatic formatter as well.

* Don't define char traits' `length()`.

This function is just not appropriate to a `std::byte` specialisation.
There is no way inspect the contents of a raw byte string and figure
out from there what the string's length is.

So, remove the function's definition.  The _declaration_ is required,
but there can't be a usable definition.