Fix compile error in `field` enum support. (#824)
Fixes #807.

Type names for user-defined enums are stored as `std::string_view`, not
as `std::string` like the others.  (Both are `inline` though.)

The problem did not affect the conversion as such, and that is why a
pure unit test was not enough to catch it.  But it did show up when
converting a `pqxx::field` value.