Only copy headers to source tree if building docs. (#856)
The `docs` target copies the header files into the build tree, so that
doxygen can find them.  But that turns out to cause trouble for regular
development: the `docs` target gets built even when not configured to
build documentation, so you're always compiling using copies of the
header files.  A real pain when editing those.

For now, just skip the copying when not configured to build the docs.
In a later iteration we can do something more sensible such as copy
the headers and the source files to a different location in the build
tree.