Rename variables from `conn` to `cx`.
I've been wanting to do this for a while.  For some reason, calling
variables and parameters `conn` (to denote a `connection`) started to
feel a bit old-fashioned and clunky.

And thinking about it some more, calling them `cx` is exactly as
valid as calling transactions `tx`.  So it's not just more compact,
it's also more consistent.