Clean up naming of `conn` & `trans` variables. (#861)
For years now I've been using `tx` as the standard name for a variable (or parameter) that denotes a transaction.  It used to be `trans`, but I liked the shorter `tx`.

But for some reason I stuck with `conn` for a connection.  Here I standardise on `cx` and `tx`.  Nice and short, but I think still clear.