Reduce string copying in array parsing.
This should reduce the copying of strings in array parsing a bit.
Parsing an unquoted string is really just a matter of referring to the
input data, so we can use a `std::string_view` there instead of a
`std::string`.

Upon reflection I don't think this breaks the ABI, since it's purely an
internal call.  _Yes_ there's an externally visible symbol involved but
nobody has any business calling it.