Work around clang compile problem in a test.
For some reason the line in this unit test where we create a
`std::vector<pqxx::binarystring>` triggers a deprecation warning, even
though the line is bracketed in an "ignore deprecations" block.

Since the whole class is deprecated, it wasn't really worth fixing
properly.  I'll just stop building this test on clang, until the time
comes to drop the whole class and its tests anyway.  Which I hope is
not too far off.