Drop assertion in workaround for `std::unreachable`.
In environments where `std::unreachable` is not available, we previously
_asserted_ that the code was not reached.  Which I guess was useful in
the early days for exposing any mixups, but doesn't really set the same
expectations as `std::unreachable` itself.

So, drop that assertion.  This also gets rid of a bunch of warnings
from my static checker about "this could be a `static_assert`."