Fix string conversion for nulls in arrays. (#907)
Fixes: #906

The budget calculation for the conversion from array-like types to SQL strings did not take null values into account. (In a distant past, SQL arrays could not contain null values.) The actual string conversion did generate a null, but the budget calculation simply tried to generate a number — which could be less than what was needed to insert the SQL NULL keyword.

Next, the budgeting function for std::optional failed to account for the possibility of the std::optional containing a value, but the value being a null.

Also the buffer budget calculation for arrays of unquoted-safe types was slightly too tight. I think I've had to fix this for unsafe types in the past.

(I'm also removing an internal wrapper function for comparing std::string_view values for equality. That type supports the equality operator nowadays.)