Replace `errorhandler` mechanism. (#903)
Deprecating `errorhandler`.  The whole mechanism is outdated and clunky
and also horribly _broken_ when a handler survives its `connection`.
Turns out `PGresult` keeps a pointer to the same handler that the
`PGconn` had, so it can still send notices even after you've closed the
connection.

Trying to fix this with the existing mechanism would still break the
ABI.  (At least without infernal and probably non-portable trickery.)

Instead of modernising the whole mechanism, I'm also simplifying it.
I doubt many people will need _chains_ of handlers, with the option
for handlers to request bypassing of the following handlers.  And for
those rare cases where that's useful, it's probably easier for the
application to implement the call chain itself!