As per @KayEss's suggestion: `notification` struct.
Yes, that's better.

Still a few details to be ironed out: Should we pass the `notification`
by value, by `const` reference, or by rvalue reference?  There's things
to be said for each option, but I like the simplicity of passing by
value.  Ultimately I doubt it matters much.

Another one is the name for the `pqxx::connection &` field.  We can't
call it `connection` because that's a type name.  (Probably technically
allowed but some compilers will complain.)  I've been using "cx" as the
standard abbreviation for "connection" — but I don't really feel like
breaking consistency with other class members today.  Perhaps it makes
sense to settle on "conn" for class members and "cx" for local
variables.