Retire result row slicing.
This was a long-forgotten feature that somebody thought might be a
good idea, around 20 years ago.  It's only been deprecated for a
short time, but it's so obscure that I stronly doubt that anyone has
ever used it at all.

Meanwhile it was weighing down a pretty basic piece of code that
comes up in lots of loops, so... less is more.