Make deprecated `field` ctors public.
These constructors have been documented as "do not use" in various ways
for a long time.

More generally, I'm settling on a deprecation horizon of 3 years.  It's
been so long since 7.0 came out that it's just not reasonable to keep
everything that was deprecated since then.  Instead, I'm now retiring
things that were deprecated in 7.6.