Use `__cpp_has_attribute()` (#928).
This changes the feature checks fo attributes `[[assume]]`,
`[[likely]]`, and `[[unlikely]]` to rely on the C++20 standard means of
testing for attributes.

I may want to systematise this into a separate mechanism, so that there
will be just a list of attrbutes that the configure process needs to
check for.