Assume support for `std::source_location`. (#941)
* Assume support for `std::source_location`.

This is a C++20 feature, and it seems to be widely implemented.

Next step of course is to pass the source location in more places, so
that exceptions will be properly helpful in pinpointing where their
errors happened.  Ideally an error will tell you the location of the
original call into libpqxx that led to the problem.

* Use `std::map::contains()`.