Pass `std::source_location` in array code.
This is just one small portion of the work needed.  I'm trying to ensure
that every libpqxx function that may throw an exception has a
`source_location` parameter, and that it passes this to any functions
it calls that may throw.

The end result would ideally be that any exception thrown by libpqxx
would include the location of the original call into libpqxx.

In practice unfortunately I'm not so sure that I can attain this
ideal.  Type parameter packs make it difficult.