Pass `std::source_location` in `field` code.
There's a lot of places where I could not do this yet because of
templating.  Frankly I feel like the whole `std::source_location` is
waiting for a do-over.

But it's hard to imagine how it should work...  Maybe by having some
way to retrieve the call site from _within_ the callee.  That would
eliminate the gruesome parameter-passing.