Move `encoding_group` into the public API.
The new string conversion API will include encoding information.  We
need that for array/composite parsing.

I was thinking to hide this away in a class, but...  it's starting to
sound like too much overhead, both for the hardware and for humans.
We've had this "encoding groups" system for years now, and haven't
seen much need to change it.

There are _some_ changes: I just added an `UNKNOWN` value to the enum
so we can deal at least somewhat gracefully with cases where we don't
have the information.  And I'm looking forward in the 8.0 release to
retiring a few of the encoding groups.  Perhaps there are ones we'll
want to add as well, but adding is cheap.

Compilers and tooling will warn about unhandled cases in a `switch`,
and that gives me more confidence that people won't be surprised by
unexpected new groups.  At least, nobody who cares about keeping their
code correct, enables compiler warnings, and pays attention.