Fix some CodeQL "passing large object" warnings.
The code scan on GitHub was complaining about some slightly large
objects were being passed as arguments in test code.

Not a huge deal, since this is test code and we're not talking about
stack-threatening amounts, but fixing is better than ignoring.

There were a few others that are a bit harder to fix, and which may be
gone anyway in the current 8.0 code.