#961: Introduce `copy_chars()` helper. (#969)
Introduce `copy_chars()` helper.

Fixes: https://github.com/jtv/libpqxx/issues/961

This is something I kept needing: a bounds-checking version of
`std::string_view::copy()`, which writes into a `std::span<char>`
instead of a raw pointer; which just copies the entire string without
needing an argument to say so; and which accepts an offset into the
destination buffer (not one into the source),

As I upgrade more string conversions to the libpqxx 8.0 API, I expect to
need this in many more places.