#959: Replace `type_name` template with `name_type()`. (#970)
Fixes: #959

This should make it easier to override a type name with a manually
written, human-friendly one such as std::string instead of the whole
std::basic_string<char, ...> incantation; inline those to reduce
duplication; and consistently make it a std::string_view.

With that latter point, it abstracts away the underlying storage. Which does mean that sometimes with template types you may need to construct the names in a separate place just so name_type() can return a view. But that's probably better than storing each individual name in its own std::string, and allows for more optimisations.