Fix a bug `get_target_property(PostgreSQL_INCLUDE_DIRS PostgreSQL::PostgreSQL)` does not working
`get_target_property(PostgreSQL_INCLUDE_DIRS PostgreSQL::PostgreSQL INTERFACE_INCLUDE_DIRECTORIES)` is not guaranteed to work.

It is not mentioned anywhere in the documentation as a property of the `PostgreSQL::PostgreSQL` target.
`PostgreSQL_INCLUDE_DIRS` is a variable automatically defined by
`find_package(PostgreSQL)`.

It should be removed because if it cannot be retrieved for a property,
the build will fail.