Make `always_null` default to `false`.
I hate setting defaults for things like this.  But it _is_ the "safe"
value, not just the "normal" value.

The reason for setting it is that the compiler error for "there is no
`pqxx::string_traits` specialisation for this type" often seemed to say
that the type's `always_null` was not `constexpr`.  That's just too
confusing.