Change (and _specify!)_ meaning of comparisons.
This drastically changes the meaning of the comparison operators on
`result`, and `row`.  (With `row_ref` and `field_ref` getting the new
semantics.)

For `field` I had actually specified semantics, so I'm not going to
change that quite as aggressively: I merely marked that as deprecated
so I can change its meaning in 9.0.