987: Deprecate `ignore_unused()`.
Turns out we haven't needed this helper for a long time!

Standard tools for the same thing:
* `std::ignore` (C++11)
* `[[maybe_unused]]` (C++17)
* `_` (C++26?)

So I've removed all uses of `pqxx::ignore_unused()`, marked the
function deprecated, and —irony of ironies— marked the thing
`[[maybe_unused]]`.

It's a bit late to find this out, but while searching for information on
this topic I discovered that Boost _also_ defined a function for this,
which had the exact same name.  Form follows function, I suppose!