Replace a use of inefficient `get_glyph_scanner()`.
This gets rid of 1 of 2 uses of that slow, naïve text scanning method.
Instead of a _glyph scanner,_ we want to use a _char finder_ in all
cases that we can.  That's much more efficient, because it inlines the
scanning (my term for finding the end of a character in a string) into
a bigger loop.

If I can get rid of the other use-case as well, not only does it
become possible to ditch that code, but also we'll be able to use the
monobyte encoding group for all "ASCII-safe" encodings.  (By which I
mean encodings where no byte inside a multibyte character ever has the
same value as an ASCII character.)  We simply won't need an encoding
group for e.g. UTF-8, because it'll effectively be a monobyte encoding.
All scanning will be done in the form of finding ASCII characters.

The new loop is pretty clever.  Which is not normally a good thing.
That's why I beefed up the tests — Kernighan's Law.  I hope it's also
very efficient, with minimal conditionals or special cases (once the
loop hits its stride).