Eliminate last use of `for_glyphs()`.
This will speed up text parsing in this particular situation (query
text when constructing a cursor), and it makes some code obsolete.

Again I wrote a bunch of compile-time unit tests for a local function
that the regular tests wouldn't be able to call.

Had to stop using `std::isspace()` because it's non-`constexpr` and it
was slightly too complicated for my taste.  I'm sure there are some
obscure ASCII space characters that I forgot to check.