Fix conversion from text to binary.
The conversion stupidly still used, and returned, a buffer of type
`pqxx::bytes`.  That's far too specific — it's meant to convert to any
type suitable for binary data.

There's a small complication: the conversion calls `resize(std::size_t)`
on the buffer, which won't work on types that don't have that member
function.  However, it's also possible that the caller knows the right
size at compile time and uses a type like `std::array` of that number of
bytes.  So if there's no `resize()` method but the buffer has the right
size from the start, that's fine too.