Don't bother with `npos` when scanning characters.
I reckon since we'll always have the buffer size in a register when we
do that check, it's probably going to be slightly easier on the CPU to
return that value, rather than a constant.  The compiler's register
allocation can map the two uses to the same register, so it could be
really trivial at run time.  Depending on the architecture, it may
actually be harder to load a constant.  The caller will check against
the end of the buffer anyway.

And while we're at it... why should we check against the end of the
buffer both inside the scanning functions and in the caller?  If we
can make it part of the contract that we never call the scanning
functions beyond the end of the buffer, the scanning functions
themselves won't need to check.  Perhaps that's a bit redundant
though, since all this stuff will get inlined eventually and the
compiler should be able to eliminate the double check.